fix(config-ui): adjust the connection request
